Give Velocity time and cost cards one template
The cost card put four large numbers in a 2x2 grid. The numbers did not
measure the same thing, and the wrapped hint text made the grid ragged.
The time card used a different shape, so the two cards in the same row
did not match in height and the left card ended in empty space.

Both cards now use one template: a total, a labeled split with color
dots, a short note, and a chart. The dots name the bands in the chart,
so neither chart needs a legend. The time split shows no percentages,
because a median of sums is not the sum of medians.

The page also passed the ESLint max-lines budget, which blocked CI. The
mock data and the charts move to separate modules.
